Factor out common utilities for stochastic tests (#10108)
This extracts some of the core utilities used in writing SharedTreeFuzzTests.ts into re-usable portions, putting them into a different folder. As part of this work, I refactored the Id compressor fuzz tests to take a similar approach. These utilities model a stochastic test as a series of serializable Operations that are applied to some state. Operations get generated using a source of randomness (provided via random-js), and applied via a custom application function. Merge-tree's operation runner is already structured generally like this, though like the previous IdCompressor fuzz test the state object is implicit. (Unlike IdCompressor, it does semantically separate generation of an operation from application of it).

These utilities should help streamline the process of running the stochastic test, and can aid in getting consistent developer experience (e.g. "write to file on failure") without having to rewrite that functionality for each system we want to test. Generally, creating a new stochastic test suite should entail defining some state interface for the test, and a discriminated union of operations that can be performed on that state. Then, a stream of those operations needs to be generated in some fashion, which is what all of the generator-creating helpers are useful for. Different domains may have different "interesting" examples of fuzz tests, e.g. SharedTreeFuzzTests's "upgrading halfway through" test case.

Co-authored-by: Abram Sanderson <absander@microsoft.com>