Allow Quorum propose() promises to remain unsettled until reconnect after disconnect (#10213)
When we get disconnected, it's possible our outstanding proposal actually was sequenced and we just didn't see the ack before disconnecting. In this case, we'll see the ack on reconnect and hit the 0x1d1 assert.

This change restructures the handling of local proposal promises and permits them to remain unsettled until the reconnect completes -- this is when we learn whether our outstanding proposals were seen by the server or got vanished.